Improved performance
The massive changes on file here basically serve one goal: to improve
performance of the driver. Along with an upstream change to mint, the
fixes here ensure that erlang's binary match context optimizations are
applied when we parse data from the server. This means allocations are
drastically reduced, and we see a roughly 4x performance improvement
over the mint-sync branch, which itself is 3x faster than master.

Making use of the binary optimizations is very macro-prone, as they're
only applied in a small set of circumstances, and would be almost
impossible to make happen without macros.